Deepa is a very smart lady (fix int vs double issue)

+++ getNS2Data.m
@@ -23,7 +23,7 @@ fprintf('Found %d channels of NS2 data.\n',hdr2.hdr.nChans);
 for tind = 1:length(dat)
     epochStartTime = dat(tind).time(1) - nsEpoch(1);
     epochEndTime = dat(tind).time(2) + nsEpoch(2);
-    nsEndTime = hdr2.hdr.nSamples / hdr2.hdr.Fs;
+    nsEndTime = double(hdr2.hdr.nSamples) / double(hdr2.hdr.Fs);
     if epochStartTime < 0
         epochStartTime = 0;
     end
